Renamed push notification manager. Added auth task listener
